added functionality of deleting lists locally and globally if the owner deletes it.

+++ app/src/main/java/com/example/tomer/shopartment/activities/MainActivity.java
@@ -37,6 +37,7 @@ import com.example.tomer.shopartment.holders.ItemViewHolder;
 import com.example.tomer.shopartment.models.Invite;
 import com.example.tomer.shopartment.models.Item;
 import com.example.tomer.shopartment.models.ShoppingList;
+import com.example.tomer.shopartment.models.User;
 import com.firebase.ui.firestore.FirestoreRecyclerAdapter;
 import com.firebase.ui.firestore.FirestoreRecyclerOptions;
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -74,7 +75,6 @@ public class MainActivity extends AppCompatActivity {
     ImageButton addItemButton;
     ArrayList<Item> createdItems;
     ArrayList<ShoppingList> userLists;
-    ChooseListAdapter chooseListAdapter;
     ImageView profilePic;
     TextView username;
     TextView email;
@@ -95,6 +95,7 @@ public class MainActivity extends AppCompatActivity {
     String currListId;
     CollectionReference currentListRef;
     CollectionReference listToDeleteRef;
+    CollectionReference sharedByRef;
     ShoppingList currentListModel = null;
     FragmentManager fragmentManager;
 
@@ -112,6 +113,7 @@ public class MainActivity extends AppCompatActivity {
         userRef = firestoreDB.collection("users").document(mAuth.getCurrentUser().getEmail());
         userShoppingListRef = firestoreDB.collection("lists").document(mAuth.getCurrentUser().getEmail()).collection("userLists");
 
+
         // set views
         searchbar =  findViewById(R.id.searchbar_edit_text);
         addItemButton = findViewById(R.id.searchbar_plus_icon);
@@ -123,7 +125,6 @@ public class MainActivity extends AppCompatActivity {
         userLists = new ArrayList<>();
 
         // set adapters
-        chooseListAdapter = new ChooseListAdapter(userLists);
         itemRecyclerView = findViewById(R.id.itemsRecycleView);
         RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(MainActivity.this);
         itemRecyclerView.setLayoutManager(layoutManager);
@@ -165,27 +166,75 @@ public class MainActivity extends AppCompatActivity {
         }
         if(!currentListModel.getCreatedBy().equals(mAuth.getCurrentUser().getEmail())) {
             // delete only localy, because its not the current user's list
+
             firestoreDB.collection("lists")
                     .document(mAuth.getCurrentUser().getEmail())
                     .collection("userLists")
                     .document(currentListModel.getId()).delete();
+            if (sharedByRef != null){
+                sharedByRef.document(mAuth.getCurrentUser().getEmail()).delete();
+            }
+            sharedByRef = null;
             currentListModel = null;
             currentListRef = null;
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
             setTitle("Shopa");
 
         }
-        // current user created the list, need to delete for all the users sharing
-        listToDeleteRef = currentListRef;
-        clearList();
-        firestoreDB.collection("lists")
-                .document(mAuth.getCurrentUser().getEmail())
-                .collection("userLists")
-                .document(currentListModel.getId()).delete();
-        currentListModel = null;
-        currentListRef = null;
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        setTitle("Shopa");
+        else {
+            // current user created the list, need to delete for all the users sharing
+            listToDeleteRef = currentListRef;
+            // currentListModel.setDeleted(true);
+            deleteForAll();
+
+        }
+
+    }
+
+    private void deleteForAll() {
+        firestoreDB.collection("items")
+                .document(currentListModel.getId())
+                .collection("sharedBy")
+                .get()
+                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (task.isSuccessful()) {
+                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                String userEmail = (String) document.get("email");
+                                currentListModel.setDeleted(true);
+                                firestoreDB.collection("lists")
+                                        .document(userEmail)
+                                        .collection("userLists")
+                                        .document(currentListModel.getId())
+                                        .set(currentListModel);
+                                firestoreDB.collection("items")
+                                        .document(currentListModel.getId())
+                                        .collection("sharedBy")
+                                        .document(userEmail).delete();
+                                Log.d(TAG, document.getId() + " => " + document.getData());
+                            }
+
+                            clearList();
+                            firestoreDB.collection("lists")
+                                    .document(mAuth.getCurrentUser().getEmail())
+                                    .collection("userLists")
+                                    .document(currentListModel.getId()).delete();
+                            firestoreDB.collection("items")
+                                    .document(currentListModel.getId())
+                                    .delete();
+                            sharedByRef = null;
+                            currentListModel = null;
+                            currentListRef = null;
+                            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+                            setTitle("Shopa");
+
+                        } else {
+                            Log.d(TAG, "Error getting documents: ", task.getException());
+                        }
+                    }
+                });
+
     }
 
     public void configureAddButton() {
@@ -205,7 +254,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 vibe.vibrate(50);
-                if(currListName.equals(DEFAULT_LIST_NAME)){
+                if(currentListModel == null){
                     Toast.makeText(MainActivity.this, "Please Choose a list first!", Toast.LENGTH_LONG).show();
                     return;
                 }
@@ -251,6 +300,10 @@ public class MainActivity extends AppCompatActivity {
             int id = item.getItemId();
             switch (id) {
                 case R.id.Clear:  // clear current list (assuming we have only one list at a time).
+                    if(!currentListModel.getCreatedBy().equals(mAuth.getCurrentUser().getEmail())){
+                        Toast.makeText(MainActivity.this, "You can't clear a list you haven't created.", Toast.LENGTH_LONG).show();
+                        break;
+                    }
                     listToDeleteRef = currentListRef;
                     clearList();
                     drawerLayout.closeDrawers();
@@ -265,9 +318,9 @@ public class MainActivity extends AppCompatActivity {
                     // open choose list fragment with all of the user's lists
                     drawerLayout.closeDrawers();
                     ChooseListFragment chooseListFragment = new ChooseListFragment();
-                    Bundle bundle = new Bundle();
-                    bundle.putParcelableArrayList("userLists" , userLists);
-                    chooseListFragment.setArguments(bundle);
+//                    Bundle bundle = new Bundle();
+//                    bundle.putParcelableArrayList("userLists" , userLists);
+//                    chooseListFragment.setArguments(bundle);
                     getSupportFragmentManager().beginTransaction().add(R.id.drawer , chooseListFragment , "frag1").addToBackStack("what").commit();
                     break;
                 case R.id.Create:
@@ -372,6 +425,7 @@ public class MainActivity extends AppCompatActivity {
 
 
     private  void clearList(){ // delete all of the items collection documents
+
         final WriteBatch deleteBatch = firestoreDB.batch();
         listToDeleteRef.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
             @Override
@@ -415,6 +469,8 @@ public class MainActivity extends AppCompatActivity {
         currListName = shoppingList.getName();
         currentListModel = shoppingList;
         currentListRef = firestoreDB.collection("items").document(shoppingList.getId()).collection("listItems");
+
+       // setListListener();
         updateScreen();
         totalPriceSet();
         setNewTitle();
@@ -642,8 +698,18 @@ public class MainActivity extends AppCompatActivity {
 
                 //when accept pressed, add shared list to user's lists and delete the invite.
                 DocumentReference inviteRef = firestoreDB.collection("invites").document(mAuth.getCurrentUser().getEmail()).collection("userInvites").document(invite.getInviteId());
-                userShoppingListRef.document(invite.getInviteId()).set(invite.getShoppingList());
+                ShoppingList shoppingList = invite.getShoppingList();
+                userShoppingListRef.document(invite.getInviteId()).set(shoppingList);
                 inviteRef.delete();
+                sharedByRef = firestoreDB.collection("items").document(shoppingList.getId()).collection("sharedBy");
+                if(!mAuth.getCurrentUser().getEmail().equals(shoppingList.getCreatedBy())){
+                    User user = new User(mAuth.getCurrentUser().getUid() ,
+                            mAuth.getCurrentUser().getDisplayName() ,
+                            mAuth.getCurrentUser().getEmail());
+                    sharedByRef.document(mAuth.getCurrentUser().getEmail()).set(user);
+                }
+                setCurrentListRef(shoppingList);
+                setListListener();
             }
         });
         builder.setNegativeButton("Decline", new DialogInterface.OnClickListener() {
@@ -667,8 +733,123 @@ public class MainActivity extends AppCompatActivity {
         return true;
     }
 
-    // TODO search for firestore offline capabilities
+    private void setListListener(){
+        final DocumentReference docRef = firestoreDB.collection("lists")
+                .document(mAuth.getCurrentUser().getEmail())
+                .collection("userLists")
+                .document(currentListModel.getId());
+        docRef.addSnapshotListener(new EventListener<DocumentSnapshot>() {
+            @Override
+            public void onEvent(@Nullable DocumentSnapshot snapshot,
+                                @Nullable FirebaseFirestoreException e) {
+                if (e != null) {
+                    Log.w(TAG, "Listen failed.", e);
+                    return;
+                }
 
+                if (snapshot != null && snapshot.exists()) {
+                    if((Boolean) snapshot.get("deleted")){
+                        String id = (String)snapshot.get("id");
+                        String name = (String)snapshot.get("name");
+                        checkIfDeleted( id , name);
+
+                    }
+
+                } else {
+                    Log.d(TAG, "Current data: null");
+                }
+            }
+        });
+//        firestoreDB.collection("lists")
+//                .document(mAuth.getCurrentUser().getEmail())
+//                .collection("userLists")
+//                .addSnapshotListener(new EventListener<QuerySnapshot>() {
+//                    @Override
+//                    public void onEvent(@Nullable QuerySnapshot value,
+//                                        @Nullable FirebaseFirestoreException e) {
+//                        if (e != null) {
+//                            Log.w(TAG, "Listen failed.", e);
+//                            return;
+//                        }
+//                        // listen success, listing all invitations on the log
+//                        if(!value.isEmpty()) {
+//                            for (DocumentSnapshot doc : value) {
+//                                ShoppingList shoppingList = (ShoppingList) doc.toObject(ShoppingList.class);
+//                                checkIfDeleted(shoppingList);
+//                            }
+//                        }
+//                    }
+//                });
+    }
+
+    private void checkIfDeleted(String shoppingListId , String shoppingListName){
+            if(shoppingListId.equals(currentListModel.getId())){
+                currentListModel = null;
+                currentListRef = null;
+                currListName = DEFAULT_LIST_NAME;
+                getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+                setTitle("Shopa");
+            }
+            firestoreDB.collection("lists")
+                    .document(mAuth.getCurrentUser().getEmail())
+                    .collection("userLists")
+                    .document(shoppingListId)
+                    .delete();
+            String msg = "Owner deleted the list: " + shoppingListName + "you won't be able to access it";
+            Toast.makeText(this, msg  , Toast.LENGTH_SHORT).show();
+    }
+
+//    private void deleteListGlobaly(){
+//        ArrayList<String> sharedBy = currentListModel.getSharedBy();
+//        for(String userEmail: sharedBy){
+//            firestoreDB.collection("lists")
+//                    .document(userEmail)
+//                    .collection("userLists")
+//                    .document(currentListModel.getId()).delete();
+//        }
+//        firestoreDB.collection("lists")
+//                .document(mAuth.getCurrentUser().getEmail())
+//                .collection("userLists")
+//                .document(currentListModel.getId()).delete();
+//        currentListModel = null;
+//        currentListRef = null;
+//        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+//        setTitle("Shopa");
+//    }
+
+//    private void setListListener(){
+//        final DocumentReference docRef = firestoreDB.collection("lists")
+//                .document(mAuth.getCurrentUser().getEmail())
+//                .collection("userLists").document(currentListModel.getId());
+//        docRef.addSnapshotListener(new EventListener<DocumentSnapshot>() {
+//            @Override
+//            public void onEvent(@Nullable DocumentSnapshot snapshot,
+//                                @Nullable FirebaseFirestoreException e) {
+//                if (e != null) {
+//                    Log.w(TAG, "Listen failed.", e);
+//                    return;
+//                }
+//                // list info changed. probably for deletion
+//                if (snapshot != null && snapshot.exists()) {
+//                    ShoppingList shoppingList = (ShoppingList) snapshot.toObject(ShoppingList.class);
+//                    if(shoppingList.getDeleted()){
+//                        docRef.delete();
+//                        currentListModel = null;
+//                        currentListRef = null;
+//                        currListName = DEFAULT_LIST_NAME;
+//                        setNewTitle();
+//                        Toast.makeText(MainActivity.this, "list got deleted by creator.", Toast.LENGTH_SHORT).show();
+//                    }
+//                } else {
+//                    Log.d(TAG, "Current data: null");
+//                }
+//            }
+//        });
+//    }
+
+
+
+    // TODO search for firestore offline capabilities
 }
